Warning fix test #4

+++ flockers/src/main.rs
@@ -47,7 +47,7 @@ fn main() {
     let state = Flocker::new(dim, num_agents);
     Visualization::default()
         .with_window_dimensions(1000., 700.)
-        .with_simulation_dimensions(dim.0 as f32, dim.1 as f32)
+        .with_simulation_dimensions(dim.0, dim.1)
         .with_background_color(Color::rgb(0., 0., 0.))
         .with_name("Flockers")
         .start::<VisState, Flocker>(VisState, state);
